requirements.txt
fastapi
uvicorn
openai
elevenlabs
sqlite3
python-dotenv
